Fix compiler warnings and add stability configuration
## Code Improvements:

### 1. Add @OptIn for ExperimentalCoroutinesApi (✅ 4 warnings fixed)
- `ImageViewModel.kt`: Added @OptIn annotation for flatMapLatest/mapLatest
- `PhotosViewModel.kt`: Added @OptIn annotation for mapLatest

### 2. Fix deprecated kotlinx.datetime.Instant usage (✅ 2 warnings fixed)
- `PlatformDatePicker.kt`: Removed import, use fully qualified name
- `RoverDateUtil.kt`: Removed import, use fully qualified name
- Both files now use `kotlinx.datetime.Instant.fromEpochMilliseconds()`

### 3. Remove redundant type conversion (✅ 1 warning fixed)
- `MissionInfoUtils.kt`: Removed unnecessary `.toLong()` call
- `abs()` already returns Long for Long operands

### 4. Create Compose stability configuration (✅ 1 warning fixed)
- Added `shared/compose-stability-config.txt`
- Marks immutable collections and domain models as stable
- Improves Compose recomposition performance

## Build Status:
✅ All platforms build successfully
✅ Android: assembleDebug
✅ iOS: linkDebugFrameworkIosSimulatorArm64
✅ Desktop: build

## Remaining Warnings:
- expect/actual Beta warnings (informational only)
- AGP compatibility (will be fixed in AGP 9.0)

+++ shared/src/commonMain/kotlin/com/sirelon/marsroverphotos/utils/MissionInfoUtils.kt
@@ -21,7 +21,7 @@ internal fun calculateEarthDaysActive(landingDate: String, maxDate: String): Lon
         val landingEpoch = landing.toEpochDays()
         val maxEpoch = max.toEpochDays()
 
-        (abs(maxEpoch - landingEpoch) + 1).toLong()
+        abs(maxEpoch - landingEpoch) + 1
     } catch (e: Exception) {
         Logger.e("MissionInfoUtils", e) { "Error calculating Earth days active" }
         0L
